requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pillow
python-decouple
dj-config-url
drf-yasg
django-filter
celery
redis
celery[redis]
django-allauth


dwebsocket
